dmioscope: add DmstatsException class
Add a generic dmstats exception class: this represents any type
of problem executing a dmstats command (command not found or not
executable, bad parameters etc.).

Functions that throw DmstatsException should log a descriptive
error message with log_error(), as well as command output with
log_verbose() to allow the detailed cause of the error to be
inspected.