dmioscope: ignore non-ioscope managed regions
Reports are not filtered by the set of managed region_id values:
this means that regions created by other programs, or the user,
may appear: skip over these regions and do not attempt to use the
returned values.

With the use of --programid when calling dmstats it is possible to
filter many of these out, however the script should still behave
correctly even if they are present.