dmioscope: correct interval duration before sleeping
Python lacks (packaged) modules to access Linux timerfd and POSIX
interval timers. The majority of the available schedulers (twisted
is one exception) seem to offer a "call function then wait for N"
model, rather than a true interval timer.

Twisted does implement a usable timer but pulls in a lot of baggage
with it; instead correct the interval duration by taking the modulo
of the time difference since the start of the run, and subtracting
any remainder from the configured interval.