dmioscope: add DmStatsCounters class
Add a class to represent the set of dmstats counters selected for
a particular histogram. The class is able to format the counter
selection as a string for display to the user ("READS,WRITES", etc.),
or as a comma-separated list of dmstats field names ("reads_count,
writes_count", etc.) to pass to 'dmstats report', and to accept a
line of data, extract the region identifier, and return the sum of
the present counter fields.

This will allow command-line counter lists to be parsed into a form
that easily allows extraction of the counter data when updating a
IOHistogram object.

Related: #5.