dmioscope: add CSV file logging backend
Add a CSVLogger class to write comma separated variable formatted
data.

The format of individual data lines is:

<iteration>,<timestamp>,<data1, data2, ... <dataN>

An initial header line (with zero interval and timestamp)

<iteration>,<timestamp>,0 * nr_bins,<width1, width2, ... <widthN>

Whenever bin widths, or the number of bins, changes, new widths
will be output in columns adjacent to that interval's data items,
in the same format as the header line.

Related: #9.