dmioscope: move Logger newline handling
The base Logger class implemented a dumb, "here's a line" IO model.

This works well for CVS, which is entirely line-oriented, but is
more tricky with JSON where we are writing (possibly) multi-line
strings, and attempting to interleave our own markup with them.

Remove the automatic linebreak insertion from Logger._write() and
allow Logger implementations to do their own thing, suffixing the
data to be written with a newline or not, as the class decides.