dmioscope: emit correct JSON when histogram bounds change
The JSONLogger formatted ouput as a succession of JSON objects in
a single file stream, for e.g.:

 {'foo': 'bar'}
 {'baz': 'qux'}

This is legal (in that each entry is itself a valid JSON object),
but most parsers expect a single completely enclosed object to
consider a stream as valid.

To address this, output an outer array that contains each of the
JTS objects. This allows an arbitrary number of objects to be
written in sequence, each with differing 'columns' and 'data'
configurations:

  [{'foo': 'bar'}, {'baz': 'qux'}]

This keeps the JSON parser happy and ensures that the right bin
descriptions are written with each JTS object.

Closes: #12.